pop: refactor Account/Mailbox/Email private data
