+++ debug/graphviz.c
@@ -49,6 +49,7 @@
 #include "notmuch/mdata.h"
 #include "notmuch/adata.h"
 #include "notmuch/private.h"
+#include "pop/adata.h"
 #include "pop/private.h"
 
 // #define GV_HIDE_CONTEXT
@@ -579,10 +580,10 @@ static void dot_mailbox_notmuch(FILE *fp, struct NmMboxData *mdata, struct ListH
   dot_object_footer(fp);
 }
 
-static void dot_mailbox_pop(FILE *fp, struct PopAccountData *mdata, struct ListHead *links)
+static void dot_mailbox_pop(FILE *fp, struct PopAccountData *adata, struct ListHead *links)
 {
-  dot_object_header(fp, mdata, "PopAccountData", "#60c060");
-  dot_ptr(fp, "conn", mdata->conn, "#ff8080");
+  dot_object_header(fp, adata, "PopAccountData", "#60c060");
+  dot_ptr(fp, "conn", adata->conn, "#ff8080");
   dot_object_footer(fp);
 }
