+++ pop/auth.c
@@ -36,6 +36,7 @@
 #include "address/lib.h"
 #include "conn/lib.h"
 #include "lib.h"
+#include "adata.h"
 #include "mutt_logging.h"
 #include "mutt_socket.h"
 #ifdef USE_SASL
